Struct extended_collections::arena::TypedArena

pub struct TypedArena<T> { /* fields omitted */ }

A fast, but limited allocator that only allocates a single type of object.

All objects inside the arena will be destroyed when the typed arena is destroyed. This typed arena also supports deallocation of objects once they are allocated and yields both mutable and immutable references to objects. Additionally, the underlying container is simply a Vec so the code itself is very simple and uses no unsafe code. When the typed arena is full, it will allocate another chunk of objects so no memory is reallocated.

Examples

use extended_collections::arena::TypedArena;

let mut arena = TypedArena::new(1024);

let x = arena.allocate(1);
assert_eq!(arena[x], 1);

arena[x] += 1;
assert_eq!(arena[x], 2);

assert_eq!(arena.free(&x), 2);

Methods

impl<T> TypedArena<T>

pub fn new(chunk_size: usize) -> Self

Constructs a new, empty TypedArena<T> with a specific number of objects per chunk.

Examples

use extended_collections::arena::TypedArena;

// creates a new TypedArena<T> that contains a maximum of 1024 u32's per chunk
let arena: TypedArena<u32> = TypedArena::new(1024);

pub fn allocate(&mut self, value: T) -> Entry

Allocates an object in the typed arena and returns an Entry. The Entry can later be used to index retrieve mutable and immutable references to the object, and dellocate the object.

Examples

use extended_collections::arena::TypedArena;

let mut arena = TypedArena::new(1024);
let x = arena.allocate(0);

pub fn free(&mut self, entry: &Entry) -> T

Deallocates an object in the typed arena and returns the object.

Panics

Panics if entry corresponds to an invalid or vacant value.

Examples

use extended_collections::arena::TypedArena;

let mut arena = TypedArena::new(1024);
let x = arena.allocate(0);
assert_eq!(arena.free(&x), 0);

pub fn get(&self, entry: &Entry) -> Option<&T>

Returns an immutable reference to an object in the typed arena. Returns None if the entry does not correspond to a valid object.

Examples

use extended_collections::arena::TypedArena;

let mut arena = TypedArena::new(1024);
let x = arena.allocate(0);
assert_eq!(arena.get(&x), Some(&0));

pub fn get_mut(&mut self, entry: &Entry) -> Option<&mut T>

Returns a mutable reference to an object in the typed arena. Returns None if the entry does not correspond to a valid object.

Examples

use extended_collections::arena::TypedArena;

let mut arena = TypedArena::new(1024);
let x = arena.allocate(0);
assert_eq!(arena.get_mut(&x), Some(&mut 0));