Module extended_collections::lsm_tree::compaction

Strategies for merging disk-resident sorted runs of data.

Structs

LeveledStrategy
A compaction strategy based on grouping SSTables into levels of exponential increasing sizes. Each SSTable is a a fixed size and two SSTables are guaranteed to be non-overlapping if they are in the same level. As smaller levels fill up, SSTables are merged into larger levels.

SizeTieredStrategy
A compaction strategy based on bucketing SSTables by their sizes and then compacting buckets when they become too full.

Traits

CompactionStrategy
Trait for types that have compaction logic for disk-resident data.

Type Definitions

CompactionIter
An iterator for the disk-resident data.