Struct extended_collections::lsm_tree::compaction::SizeTieredStrategy

pub struct SizeTieredStrategy<T, U> { /* fields omitted */ }

A compaction strategy based on bucketing SSTables by their sizes and then compacting buckets when they become too full.

Configuration Parameters

max_in_memory_size: The maximum size of the in-memory tree before it must be flushed onto disk as a SSTable.

max_sstable_count: The minimum number of SSTables in a bucket before a compaction is triggered.

min_sstable_size: The size threshold for the first bucket. All SSTables with size smaller than min_sstable_size will be bucketed into the first bucket.

bucket_low: SSTables in a bucket other than the first must have size greater than or equal to bucket_low * bucket_average where bucket_average is the average of the bucket.

bucket_high: SSTables in a bucket other than the first must have size smaller than or equal to bucket_high * bucket_average where bucket_average is the average of the bucket.

Methods

impl<T, U> SizeTieredStrategy<T, U>

pub fn new<P>(
path: P,
max_in_memory_size: u64,
max_sstable_count: usize,
min_sstable_size: u64,
bucket_low: f64,
bucket_high: f64
) -> Result<Self> where
T: Serialize,
U: Serialize,
P: AsRef<Path>,

Constructs a new SizeTieredStrategy<T, U> with specific configuration parameters.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;

let sts: SizeTieredStrategy<u32, u32> =
    SizeTieredStrategy::new("size_tiered_strategy_new", 10000, 4, 50000, 0.5, 1.5)?;

pub fn open<P>(path: P) -> Result<Self> where
T: DeserializeOwned,
U: DeserializeOwned,
P: AsRef<Path>,

Opens an existing SizeTieredStrategy<T, U> from a folder.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;

let sts: SizeTieredStrategy<u32, u32> = SizeTieredStrategy::open("size_tiered_strategy_open")?;