Trait extended_collections::lsm_tree::compaction::CompactionStrategy

pub trait CompactionStrategy<T, U> {
    fn get_path(&self) -> &Path;

    fn get_max_in_memory_size(&self) -> u64;

    fn get_and_increment_logical_time(&mut self) -> Result<u64>;

    fn try_compact(&mut self, sstable: SSTable<T, U>) -> Result<()>;

    fn flush(&mut self) -> Result<()>;

    fn get<V: ?Sized>(&mut self, key: &V) -> Result<Option<SSTableValue<U>>>
    where
        T: Borrow<V>,
        V: Ord + Hash;

    fn len_hint(&mut self) -> Result<usize>;

    fn len(&mut self) -> Result<usize>;

    fn is_empty(&mut self) -> Result<bool>;

    fn clear(&mut self) -> Result<()>;

    fn min(&mut self) -> Result<Option<T>>;

    fn max(&mut self) -> Result<Option<T>>;

    fn iter(&mut self) -> Result<Box<CompactionIter<T, U>>>;
}

Trait for types that have compaction logic for disk-resident data.

A compaction strategy should incrementally accept SSTables and handle the logic for creating new SSTables, deleting stale SSTables, and searching through current SSTables.

Required methods

fn get_path(&self) -> &Path

Returns the path of the disk-resident data.

fn get_max_in_memory_size(&self) -> u64

Returns the maximum size of the in-memory tree in bytes.

fn get_and_increment_logical_time(&mut self) -> Result<u64>

Returns and increments the current logical time of the compaction strategy.

fn try_compact(&mut self, sstable: SSTable<T, U>) -> Result<()>

Adds a SSTable to the compaction strategy and compacts the SSTables being tracked, if needed.

fn flush(&mut self) -> Result<()>

Waits until the current compaction thread, if any, terminates and updates the metadata of the compaction strategy.

fn get<V: ?Sized>(&mut self, key: &V) -> Result<Option<SSTableValue<U>>> where
T: Borrow<V>,
V: Ord + Hash,

Searches through disk-resident data and returns the value associated with a particular key. It will return None if the key does not exist in the disk-resident data.

fn len_hint(&mut self) -> Result<usize>

Returns the approximate number of items in the disk-resident data.

fn len(&mut self) -> Result<usize>

Returns the number of items in the disk-resident data.

fn is_empty(&mut self) -> Result<bool>

Returns true if the disk-resident data is empty.

fn clear(&mut self) -> Result<()>

Clears the disk-resident data, removing all values.

fn min(&mut self) -> Result<Option<T>>

Returns the minimum key of the disk-resident data. Returns None if the disk-resident data is empty.

fn max(&mut self) -> Result<Option<T>>

Returns the maximum key of the disk-resident data. Returns None if the disk-resident data is empty.

fn iter(&mut self) -> Result<Box<CompactionIter<T, U>>>

Returns an iterator over the disk-resident data. The iterator will yield key-value pairs in ascending order.

Loading content...