Struct extended_collections::lsm_tree::LsmMap

pub struct LsmMap<T, U, C> { /* fields omitted */ }

An ordered map implemented using a log structured merge-tree.

A log-structured merge-tree comprises of two components -- an in-memory tree and on-disk sorted immutable lists called Sorted Strings Tables (SSTables). The in-memory tree is incrementally flushed onto disk into SSTables when its size exceeds a certain threshold. When there are many fragmented SSTables, they are merged together using a compaction strategy. When an entry is replaced, it could occur in multiple SSTables. The value in the most recent SSTable is fetched. When an entry is deleted, a tombstone is inserted to indicate that the entry is deleted.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(0, 1)?;
map.insert(3, 4)?;

assert_eq!(map.get(&0)?, Some(1));
assert_eq!(map.get(&1)?, None);
assert_eq!(map.len()?, 2);
assert!(map.len_hint()? >= 2);

assert_eq!(map.min()?, Some(0));

map.remove(0)?;
assert_eq!(map.get(&0)?, None);

map.flush();

Methods

impl<T, U, C> LsmMap<T, U, C> where
T: Clone + Ord + Hash + DeserializeOwned + Serialize,
U: Clone + DeserializeOwned + Serialize,
C: CompactionStrategy<T, U>,

pub fn new(compaction_strategy: C) -> Self

Constructs a new LsmMap<T, U> with a specific CompactionStrategy<T, U>.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_new", 10000, 4, 50000, 0.5, 1.5)?;
let map: LsmMap<u32, u32, _> = LsmMap::new(sts);

pub fn insert(&mut self, key: T, value: U) -> Result<()>

Inserts a key-value pair into the map. If the key-value pair causes the size of the in-memory tree to exceed its size threshold, it will flush the data into a SSTable and then compact the SSTables if necessary.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_insert", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
assert_eq!(map.get(&1)?, Some(1));

map.insert(1, 2)?;
assert_eq!(map.get(&1)?, Some(2));

pub fn remove(&mut self, key: T) -> Result<()>

Removes a key-value pair into the map by inserting a tombstone. If the key-value pair causes the size of the in-memory tree to exceed its size threshold, it will flush the data into a SSTable and then compact the SSTables if necessary.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_remove", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
assert_eq!(map.get(&1)?, Some(1));

map.remove(1)?;
assert_eq!(map.get(&1)?, None);

pub fn contains_key<V: ?Sized>(&mut self, key: &V) -> Result<bool> where
T: Borrow<V>,
V: Ord + Hash,

Checks if a key exists in the map.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_contains_key", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
assert!(!map.contains_key(&0)?);
assert!(map.contains_key(&1)?);

pub fn get<V: ?Sized>(&mut self, key: &V) -> Result<Option<U>> where
T: Borrow<V>,
V: Ord + Hash,

Returns the value associated with a particular key. It will return None if the key does not exist in the map.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_get", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
assert_eq!(map.get(&0)?, None);
assert_eq!(map.get(&1)?, Some(1));

pub fn len_hint(&mut self) -> Result<usize>

Returns the approximate number of elements in the map. The length returned will always be greater than or equal to the actual length. It counts all the non-tombstone entries stored in the SSTables, so it will overcount if there are duplicate entries or if a tombstone overrides previous entries. For an accurate, but slower way of getting the length, see len.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_len_hint", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
assert!(map.len_hint()? >= 1);

pub fn len(&mut self) -> Result<usize>

Returns the number of elements in the map by first flushing the in-memory tree and then doing a full scan of all entries. For a more efficient, but approximate way of getting the length, see len_hint.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_len", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
assert_eq!(map.len()?, 1);

pub fn is_empty(&mut self) -> Result<bool>

Returns true if the map is empty. The in-memory tree is flushed and then a full scan of all entries is performed to determine if the map is empty.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_is_empty", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);
assert!(map.is_empty()?);

map.insert(1, 1)?;
assert!(!map.is_empty()?);

pub fn clear(&mut self) -> Result<()>

Clears the map, removing all values. This function will wait for any ongoing compaction thread to terminate before removing all SSTables.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_clear", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
map.insert(2, 2)?;
map.clear()?;
assert!(map.is_empty()?);

pub fn min(&mut self) -> Result<Option<T>>

Returns the minimum key of the map. Returns None if the map is empty.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_min", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
map.insert(3, 3)?;
assert_eq!(map.min()?, Some(1));

pub fn max(&mut self) -> Result<Option<T>>

Returns the maximum key of the map. Returns None if the map is empty.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_max", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
map.insert(3, 3)?;
assert_eq!(map.max()?, Some(3));

pub fn flush(&mut self) -> Result<()>

Flushes the in-memory tree into a SSTable if it is not empty. The map must be flushed before being dropped or the contents of the in-memory tree will be lost.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_flush", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
map.insert(3, 3)?;
map.flush()?;

pub fn iter(&mut self) -> Result<Box<CompactionIter<T, U>>>

Returns an iterator over the map. The iterator will yield key-value pairs in ascending order. The in-memory tree will be flushed before yielding the iterator. The map will not perform any compactions if there are any undropped iterators.

Examples

use extended_collections::lsm_tree::compaction::SizeTieredStrategy;
use extended_collections::lsm_tree::LsmMap;

let sts = SizeTieredStrategy::new("example_lsm_map_iter", 10000, 4, 50000, 0.5, 1.5)?;
let mut map = LsmMap::new(sts);

map.insert(1, 1)?;
map.insert(2, 2)?;

let mut iterator = map.iter()?.map(|value| value.unwrap());
assert_eq!(iterator.next(), Some((1, 1)));
assert_eq!(iterator.next(), Some((2, 2)));
assert_eq!(iterator.next(), None);