Struct extended_collections::skiplist::SkipList

pub struct SkipList<T> { /* fields omitted */ }

A list implemented using a skiplist.

A skiplist is a probabilistic data structure that allows for binary search tree operations by maintaining a linked hierarchy of subsequences. The first subsequence is essentially a sorted linked list of all the elements that it contains. Each successive subsequence contains approximately half the elements of the previous subsequence. Using the sparser subsequences, elements can be skipped and searching, insertion, and deletion of keys can be done in approximately logarithm time.

Each link in this skiplist store the width of the link. The width is defined as the number of bottom layer links being traversed by each of the higher layer links. This augmentation allows the list to get, remove, and insert at an arbitrary index in O(log N) time.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
list.push_back(2);
list.push_front(3);

assert_eq!(list.get(0), Some(&3));
assert_eq!(list.get(3), None);
assert_eq!(list.len(), 3);

*list.get_mut(0).unwrap() += 1;
assert_eq!(list.pop_front(), 4);
assert_eq!(list.pop_back(), 2);

Methods

impl<T> SkipList<T>

pub fn new() -> Self

Constructs a new, empty SkipList<T>.

Examples

use extended_collections::skiplist::SkipList;

let list: SkipList<u32> = SkipList::new();

pub fn insert(&mut self, index: usize, value: T)

Inserts a value into the list at a particular index, shifting elements one position to the right if needed.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
list.insert(0, 2);
assert_eq!(list.get(0), Some(&2));
assert_eq!(list.get(1), Some(&1));

pub fn remove(&mut self, index: usize) -> T

Removes a value at a particular index from the list. Returns the value at the index.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
assert_eq!(list.remove(0), 1);

pub fn push_front(&mut self, value: T)

Inserts a value at the front of the list.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.push_front(1);
list.push_front(2);
assert_eq!(list.get(0), Some(&2));

pub fn push_back(&mut self, value: T)

Inserts a value at the back of the list.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.push_back(1);
list.push_back(2);
assert_eq!(list.get(0), Some(&1));

pub fn pop_front(&mut self) -> T

Removes a value at the front of the list.

Panics

Panics if list is empty.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.push_back(1);
list.push_back(2);
assert_eq!(list.pop_front(), 1);

pub fn pop_back(&mut self) -> T

Removes a value at the back of the list.

Panics

Panics if list is empty.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.push_back(1);
list.push_back(2);
assert_eq!(list.pop_back(), 2);

pub fn get(&self, index: usize) -> Option<&T>

Returns a mutable reference to the value at a particular index. Returns None if the index is out of bounds.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
*list.get_mut(0).unwrap() = 2;
assert_eq!(list.get(0), Some(&2));

pub fn get_mut(&mut self, index: usize) -> Option<&mut T>

Returns a mutable reference to the value at a particular index. Returns None if the index is out of bounds.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
*list.get_mut(0).unwrap() = 2;
assert_eq!(list.get(0), Some(&2));

pub fn len(&self) -> usize

Returns the number of elements in the list.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
assert_eq!(list.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the list is empty.

Examples

use extended_collections::skiplist::SkipList;

let list: SkipList<u32> = SkipList::new();
assert!(list.is_empty());

pub fn clear(&mut self)

Clears the list, removing all values.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
list.insert(1, 2);
list.clear();
assert_eq!(list.is_empty(), true);

pub fn iter(&self) -> SkipListIter<T>

Returns an iterator over the list.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
list.insert(1, 2);

let mut iterator = list.iter();
assert_eq!(iterator.next(), Some(&1));
assert_eq!(iterator.next(), Some(&2));
assert_eq!(iterator.next(), None);

pub fn iter_mut(&mut self) -> SkipListIterMut<T>

Returns a mutable iterator over the list.

Examples

use extended_collections::skiplist::SkipList;

let mut list = SkipList::new();
list.insert(0, 1);
list.insert(1, 2);

for value in &mut list {
    *value += 1;
}

let mut iterator = list.iter();
assert_eq!(iterator.next(), Some(&2));
assert_eq!(iterator.next(), Some(&3));
assert_eq!(iterator.next(), None);